src/supporting-data/citations/components/results/CitationsCard.tsx
import { Card } from 'franklin-sites';
import { FC, useCallback, MouseEvent } from 'react';
import { useHistory } from 'react-router-dom';
import { SetOptional } from 'type-fest';

import LiteratureCitation from '../LiteratureCitation';

import { getEntryPath } from '../../../../app/config/urls';
import { getIdKeyFor } from '../../../../shared/utils/getIdKeyForNamespace';

import { Namespace } from '../../../../shared/types/namespaces';
import { CitationsAPIModel } from '../../adapters/citationsConverter';

const BLOCK_CLICK_ON_CARD = new Set(['A', 'INPUT', 'BUTTON']);

const getIdKey = getIdKeyFor(Namespace.citations);

const CitationCard: FC<{
  data: SetOptional<CitationsAPIModel, 'statistics'>;
  selected?: boolean;
  handleEntrySelection?: (rowId: string) => void;
}> = ({ data, selected, handleEntrySelection, children }) => {
  const history = useHistory();

  const id = getIdKey(data as CitationsAPIModel);

  const handleCardClick = useCallback(
    (event: MouseEvent) => {
      if (BLOCK_CLICK_ON_CARD.has((event.target as HTMLElement).tagName)) {
        return;
      }
      history.push(getEntryPath(Namespace.citations, id));
    },
    [history, id]
  );

  return (
    <Card onClick={handleCardClick}>
      <div className="result-card">
        {handleEntrySelection && (
          <div className="result-card__left">
            <input
              type="checkbox"
              checked={selected}
              onChange={() => handleEntrySelection(id)}
              data-testid="up-card-checkbox"
            />
          </div>
        )}
        <div className="result-card__right">
          <LiteratureCitation data={data}>{children}</LiteratureCitation>
        </div>
      </div>
    </Card>
  );
};

export default CitationCard;
